# sparta/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An Open Source Software Apply and Manager System",
        "title": "Open Source Software Manager",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/host/shutdown": {
            "post": {
                "description": "关闭指定主机",
                "tags": [
                    "主机管理"
                ],
                "summary": "关闭主机",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\t主机ip",
                        "name": "hostIP",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登陆成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/public/user/login": {
            "post": {
                "description": "登陆系统",
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登陆",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登陆成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "登陆失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user": {
            "post": {
                "description": "添加用户",
                "tags": [
                    "用户管理"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NT",
                        "name": "nt",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱地址",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "头像",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "添加成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "没有登陆",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/list": {
            "post": {
                "description": "获取用户列表",
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "没有登陆",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "description": "获取用户",
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "没有登陆",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户",
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NT",
                        "name": "nt",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "邮箱地址",
                        "name": "email",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "没有登陆",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除用户",
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "没有登陆",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}